support osc types "i" "f" & "d", use long int as cursor to run for days

+++ src/sequencer.hpp
@@ -21,7 +21,7 @@ class Sequencer
         double elapsed_samples;
 
         float bpm;
-        int cursor;
+        long cursor;
 
         bool playing;
